requirements.txt
Django==4.1.5
django_environ==0.9.0
django_filter==22.1
djangorestframework==3.14.0
drf_yasg==1.21.5
ippanel==2.0.7
jdatetime==4.1.0
python_dateutil==2.8.2
